requirements.txt
telethon
python-dotenv
selenium
webdriver-manager
pyautogui